refactor: move test create auction builder to separater folder
