+++ backend/src/Tests/FunctionalTests/Commands/Auctions/BuyNow/BuyNowCommandSuccess_Test.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
-using static FunctionalTests.Builders.TestCreateAuctionCommandBuilder;
+using static FunctionalTests.Commands.TestCreateAuctionCommandBuilder;
 using static UserPayments.DomainEvents.Events.V1;
 
 namespace FunctionalTests.Commands
