+++ backend/src/Tests/FunctionalTests/Commands/Auctions/CreateAuction/CreateAuctionCommand_Tests.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
-using static FunctionalTests.Builders.TestCreateAuctionCommandBuilder;
+using static FunctionalTests.Commands.TestCreateAuctionCommandBuilder;
 
 namespace FunctionalTests.Commands
 {
